requirements.txt
ipykernel
pandas
chardet
streamlit
streamlit_authenticator
PyYAML